Add updates to GUI and tests
