--- SegmentEditorCropSegment/SegmentEditorCropSegmentLib/__init__.py
@@ -1,4 +1,0 @@
-from SegmentEditorEffects.AbstractScriptedSegmentEditorEffect import *
-from SegmentEditorEffects.AbstractScriptedSegmentEditorLabelEffect import *
-
-from SegmentEditorEffect import *
