--- SegmentEditorCropSegment/Testing/Python/CMakeLists.txt
@@ -1,2 +1,0 @@
-
-#slicer_add_python_unittest(SCRIPT ${MODULE_NAME}ModuleTest.py)
